+++ packages/ui/src/public-api.ts
@@ -358,6 +358,8 @@ export * from './lib/menu/menu.component';
 export * from './lib/menu/menu-item.component';
 export * from './lib/menu/menu-trigger.directive';
 export * from './lib/paginator/paginator.component';
+export * from './lib/popover/popover-positioning';
+export * from './lib/popover/popover.component';
 export * from './lib/progress-bar/progress-bar.component';
 export * from './lib/radio/radio-group.component';
 export * from './lib/radio/radio.component';